refactor(comp): Rename components to comp
