+++ src/app/comp/redir/redir.component.ts
