+++ src/app/comp/xunk-spinner/xunk-spinner.component.ts
